+++ intg-denonavr/config.py
@@ -5,14 +5,16 @@ Configuration handling of the integration driver.
 :license: Mozilla Public License Version 2.0, see LICENSE for more details.
 """
 
+from collections.abc import Callable, Iterator
 import dataclasses
+from dataclasses import dataclass
+from enum import StrEnum
 import json
 import logging
-import os
-from dataclasses import dataclass
-from enum import Enum
-from typing import Iterator, Optional
+from pathlib import Path
+from typing import Any
 
+from typing_extensions import override
 from ucapi import EntityTypes
 
 _LOG = logging.getLogger(__name__)
@@ -20,7 +22,7 @@ _LOG = logging.getLogger(__name__)
 _CFG_FILENAME = "config.json"
 
 
-def create_entity_id(avr_id: str, entity_type: EntityTypes, sub_type: Optional[str] = None) -> str:
+def create_entity_id(avr_id: str, entity_type: EntityTypes, sub_type: str | None = None) -> str:
     """Create a unique entity identifier for the given receiver and entity type."""
     # Warning: only works for a single media-player entity. Doesn't work for #21 multiple zones!
     #          Zone parameter will be required.
@@ -71,7 +73,7 @@ class AvrDevice:
     is_dirac_supported: bool | None
 
 
-class SensorType(str, Enum):
+class SensorType(StrEnum):
     """Sensor types for Denon AVR."""
 
     VOLUME_DB = "volume_db"
@@ -105,7 +107,7 @@ class SensorType(str, Enum):
     DIGITAL_CODEC = "digital_codec"
 
 
-class SelectType(str, Enum):
+class SelectType(StrEnum):
     """Select types for Denon AVR."""
 
     SOUND_MODE = "sound_mode"
@@ -121,7 +123,7 @@ class SelectType(str, Enum):
     DIGITAL_CODEC = "digital_codec"
 
 
-class AdditionalEventType(str, Enum):
+class AdditionalEventType(StrEnum):
     """Additional event types for the integration."""
 
     RAW_SOUND_MODE = "RAW_SOUND_MODE"
@@ -159,8 +161,9 @@ class AdditionalEventType(str, Enum):
 class _EnhancedJSONEncoder(json.JSONEncoder):
     """Python dataclass json encoder."""
 
-    def default(self, o):
-        if dataclasses.is_dataclass(o):
+    @override
+    def default(self, o: Any) -> Any:
+        if dataclasses.is_dataclass(o) and not isinstance(o, type):
             return dataclasses.asdict(o)
         return super().default(o)
 
@@ -168,14 +171,19 @@ class _EnhancedJSONEncoder(json.JSONEncoder):
 class Devices:
     """Integration driver configuration class. Manages all configured Denon/Marantz devices."""
 
-    def __init__(self, data_path: str, add_handler, remove_handler):
+    def __init__(
+        self,
+        data_path: str,
+        add_handler: Callable[[AvrDevice], None] | None,
+        remove_handler: Callable[[AvrDevice | None], None] | None,
+    ) -> None:
         """
         Create a configuration instance for the given configuration path.
 
         :param data_path: configuration path for the configuration file and client device certificates.
         """
         self._data_path: str = data_path
-        self._cfg_file_path: str = os.path.join(data_path, _CFG_FILENAME)
+        self._cfg_file_path: Path = Path(data_path) / _CFG_FILENAME
         self._config: list[AvrDevice] = []
         self._add_handler = add_handler
         self._remove_handler = remove_handler
@@ -197,10 +205,7 @@ class Devices:
 
     def contains(self, avr_id: str) -> bool:
         """Check if there's a device with the given device identifier."""
-        for item in self._config:
-            if item.id == avr_id:
-                return True
-        return False
+        return any(item.id == avr_id for item in self._config)
 
     def add_or_update(self, avr: AvrDevice) -> None:
         """
@@ -264,8 +269,8 @@ class Devices:
         """Remove the configuration file."""
         self._config = []
 
-        if os.path.exists(self._cfg_file_path):
-            os.remove(self._cfg_file_path)
+        if self._cfg_file_path.exists():
+            self._cfg_file_path.unlink()
 
         if self._remove_handler is not None:
             self._remove_handler(None)
@@ -277,11 +282,11 @@ class Devices:
         :return: True if the configuration could be saved.
         """
         try:
-            with open(self._cfg_file_path, "w+", encoding="utf-8") as f:
+            with self._cfg_file_path.open("w+", encoding="utf-8") as f:
                 json.dump(self._config, f, ensure_ascii=False, cls=_EnhancedJSONEncoder)
             return True
         except OSError:
-            _LOG.error("Cannot write the config file")
+            _LOG.exception("Cannot write the config file")
 
         return False
 
@@ -292,7 +297,7 @@ class Devices:
         :return: True if the configuration could be loaded.
         """
         try:
-            with open(self._cfg_file_path, "r", encoding="utf-8") as f:
+            with self._cfg_file_path.open(encoding="utf-8") as f:
                 data = json.load(f)
             needs_migration = False
             for item in data:
@@ -321,11 +326,11 @@ class Devices:
                 self.store()
             return True
         except OSError:
-            _LOG.error("Cannot open the config file")
+            _LOG.exception("Cannot open the config file")
         except ValueError:
-            _LOG.error("Empty or invalid config file")
+            _LOG.exception("Empty or invalid config file")
 
         return False
 
 
-devices: Devices | None = None  # pylint: disable=C0103
+devices: Devices | None = None